M5_hello_world_program/pyproject.toml
[project]
name = "rag-hello-world"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.11.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-google-genai>=2.0.10",
    "langchain-huggingface>=0.3.0",
    "langchain-openai>=0.3.27",
    "openai>=1.93.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "tiktoken>=0.9.0",
]
